Introduce only SqlSelectListTransformer
