+++ coral-hive/src/main/java/com/linkedin/coral/hive/hive2rel/HiveToRelConverter.java
@@ -69,7 +69,7 @@ public class HiveToRelConverter extends ToRelConverter {
   }
 
   @Override
-  public SqlValidator getSqlValidator() {
+  protected SqlValidator getSqlValidator() {
     return sqlValidator;
   }
