+++ coral-service/src/main/java/com/linkedin/coral/coralservice/utils/TranslationUtils.java
@@ -25,7 +25,7 @@ public class TranslationUtils {
 
   public static String translateHiveToTrino(String query) {
     RelNode relNode = new HiveToRelConverter(hiveMetastoreClient).convertSql(query);
-    return new RelToTrinoConverter(hiveMetastoreClient).convert(relNode);
+    return new RelToTrinoConverter().convert(relNode);
   }
 
   public static String translateHiveToSpark(String query) {
