+++ coral-spark/src/main/java/com/linkedin/coral/spark/CoralSpark.java
@@ -132,10 +132,7 @@ public class CoralSpark {
     // only do this when it's not a select star case,
     // since for select star we don't need to add any explicit aliases
     if (rewritten.getKind() == SqlKind.SELECT && ((SqlSelect) rewritten).getSelectList() != null) {
-      if (!(((SqlSelect) rewritten).getSelectList().size() == 1
-          && ((SqlSelect) rewritten).getSelectList().get(0).toString().equalsIgnoreCase("*"))) {
-        rewritten = rewritten.accept(new AddExplicitAlias(aliases));
-      }
+      rewritten = rewritten.accept(new AddExplicitAlias(aliases));
     }
     return rewritten.toSqlString(SparkSqlDialect.INSTANCE).getSql();
   }
