+++ coral-trino/src/main/java/com/linkedin/coral/trino/rel2trino/RelToTrinoConverter.java
@@ -32,7 +32,6 @@ import org.apache.calcite.sql.parser.SqlParserPos;
 import org.apache.calcite.sql.type.SqlTypeName;
 
 import com.linkedin.coral.com.google.common.collect.ImmutableList;
-import com.linkedin.coral.common.HiveMetastoreClient;
 import com.linkedin.coral.common.functions.FunctionFieldReferenceOperator;
 import com.linkedin.coral.hive.hive2rel.rel.HiveUncollect;
 import com.linkedin.coral.trino.rel2trino.functions.TrinoArrayTransformFunction;
@@ -60,30 +59,18 @@ public class RelToTrinoConverter extends RelToSqlConverter {
    * For uses outside LinkedIn, just ignore this configuration.
    */
   private Map<String, Boolean> configs = new HashMap<>();
-  private HiveMetastoreClient _hiveMetastoreClient;
 
-  @Deprecated
+  /**
+   * Creates a RelToTrinoConverter.
+   */
   public RelToTrinoConverter() {
     super(TrinoSqlDialect.INSTANCE);
   }
 
-  @Deprecated
   public RelToTrinoConverter(Map<String, Boolean> configs) {
     super(TrinoSqlDialect.INSTANCE);
     checkNotNull(configs);
     this.configs = configs;
-  }
-
-  public RelToTrinoConverter(HiveMetastoreClient mscClient) {
-    super(TrinoSqlDialect.INSTANCE);
-    _hiveMetastoreClient = mscClient;
-  }
-
-  public RelToTrinoConverter(HiveMetastoreClient mscClient, Map<String, Boolean> configs) {
-    super(TrinoSqlDialect.INSTANCE);
-    checkNotNull(configs);
-    this.configs = configs;
-    _hiveMetastoreClient = mscClient;
   }
 
   /**
@@ -94,9 +81,7 @@ public class RelToTrinoConverter extends RelToSqlConverter {
   public String convert(RelNode relNode) {
     RelNode rel = convertRel(relNode, configs);
     SqlNode sqlNode = convertToSqlNode(rel);
-    SqlNode transformedSqlNode = sqlNode.accept(new SqlNodeConverter(_hiveMetastoreClient));
-
-    SqlNode sqlNodeWithUDFOperatorConverted = transformedSqlNode.accept(new CoralToTrinoSqlCallConverter(configs));
+    SqlNode sqlNodeWithUDFOperatorConverted = sqlNode.accept(new CoralToTrinoSqlCallConverter(configs));
     return sqlNodeWithUDFOperatorConverted.accept(new TrinoSqlRewriter()).toSqlString(TrinoSqlDialect.INSTANCE)
         .toString();
   }
